Update Soma.java
Sum double.

+++ Soma.java
@@ -1,9 +1,9 @@
 public class Soma{
   public static void main (String [] args){
-    int a=0, b=0, soma=0;
+    double a=0, b=0, soma=0;
     Scanner ler = new Scanner(System.in);
-    a = ler.nextInt();
-    b = ler.nextInt();
+    a = ler.nextDouble();
+    b = ler.nextDouble();
     soma = a +b;
     System.out.println("Soma:"+soma);
   }
